Add files via upload
This is an example for use of GRID's using dashed-lines to be used as overlay when creating art in processing. It helps in the design as you have an grid that can be used for positioning your components. Right-mouse click shows you the coordinates in the console.
I made this on request for someone starting Processing and finding it almost impossible to place stuff where wanted. Now the coordinates are known and layout is much easier.
The dashed-lines class can be used in any other project as well where they might be usefull, they follow the Javascript definition, based on previous work: from J David Eisenberg  https://openprocessing.org/sketch/7013/#